Add err2name and err2string methods
Unforunately the testing is a little wonky for err2str. It seems there are
other internal string representations that exist, instead of being part of
error descriptions.

I don't love the idea of creating and maintaining a complete copy of the
error enum, as well as the internal string representations of them. Probably
revisit later.