Run Travis build in Docker container
Move the remote dev files into a folder to help de-clutter the main directory.
Create a new directory travis_build to keep files specific for the Travis CI build.
Create a new Dockerfile for use in the Travis build.
Lock down the version of librdkafka being used for current development, and
update tests to reflect that.
Having problems finding the .so file in the container, so hard-code the path to
/usr/local/lib. Will need to add logic to be smarter about where to find shared libraries.